Implement support for color grading maps.
The color grader thus far supports adjustments of shadows, midtones and
highlights, with the option to preserve luminosity. It also supports standard
brightness and contrast.

Implementation wise the color grader produces a 3D texture which acts as a
look-up table for associating RGB values into color graded ones. The 3D texture
consists of 16 layers of a cube, with the y-axis mapping to green, the x-axis
mapping to red, and the z-axis mapping to blue. The identity look-up table can
be seen as (in J) "17*16 16 16#:i.4096".